requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
lightgbm>=4.0.0
h3>=3.7.6
holidays>=0.35
python-dateutil>=2.8.2

